backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
langchain
langchain-groq
python-dotenv
pytest
